Extract CombatServices from BuildMgr
Splits BuildMgr's paradigm-agnostic combat utilities into a mixin so a
second build base can share the combat vocabulary without inheriting the
generator execution engine. Pure relocation: BuildMgr 2256 -> 671 lines,
no behaviour change.

Moved (65): targeting, aggro, party health/spike monitor, spirit
placement, custom-skill inspection, auto-attack, blacklist, whiteboard,
cast API.
Kept (34): identity/matching, fallback chain, tick state, the generator
engine (_process_phase, _yield_from_handler, Process*, Tick, Set*Fn) and
BuildRegistry.

Done as an AST pass, not by hand. Verified: moved and kept method sets
partition the original 98 exactly with no overlap, every relocated body
is AST-identical to its upstream original, every self.* a moved method
touches resolves on BuildMgr, __init__ still sets all 11 service fields
in the same order, compileall passes on Py4GWCoreLib + HeroAI under
3.13, and pyflakes output is unchanged at 11 warnings.
